fix: moved the shared feature logic to leaflet.svelte to enable zooming into once the map has loaded
