+++ src/routes/+page.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-  import questionMarkSVG from "$assets/material-icons/question_mark.svg?raw";
-
   import Leaflet from "$components/map/Leaflet.svelte";
   import ZoomControl from "$components/map/ZoomControl.svelte";
   import MapControl from "$components/map/MapControl.svelte";
@@ -8,23 +6,16 @@
   import SearchOptions from "$components/inputs/SearchOptions.svelte";
   import Sidebar from "$components/sidebars/Sidebar.svelte";
   import Drawer from "$components/sidebars/Drawer.svelte";
-  import SvgIcon from "$components/icons/SVGIcon.svelte";
 
   import { allFeatures } from "$lib/stores/mapState.svelte";
 
   import { type PageData } from "./$types";
-  import type { Feature } from "$lib/types/features";
   import { getAppState } from "$lib/stores/appState.svelte";
   import { getMapState } from "$lib/stores/mapState.svelte";
-  import { getSearchState } from "$lib/stores/SearchState.svelte";
   import { onMount } from "svelte";
-  import { page } from "$app/state";
 
   let { data }: { data: PageData } = $props();
-  let id = page.url.searchParams.get("id") || "";
-  let dialog: HTMLDialogElement | undefined;
   const mapState = getMapState();
-  const searchState = getSearchState();
   const appState = getAppState();
 
   onMount(() => {
@@ -38,22 +29,7 @@
       ...data.parking!.features,
       ...data.pointsOfInterest!.features,
     ]);
-
-    if (id && $allFeatures) {
-      const feature = $allFeatures.find((f: Feature) => f.id === id);
-      if (feature) {
-        searchState.updateDetailedFeature(feature);
-        searchState.updateQuery(feature.properties.name);
-      } else {
-        dialog?.showModal();
-      }
-    }
   });
-
-  function closeDialog() {
-    dialog?.close();
-    appState.collapsedSidebar = false;
-  }
 </script>
 
 <div class="viewport">
@@ -81,14 +57,6 @@
     {/if}
   </Leaflet>
 </div>
-
-<dialog class="error-message" bind:this={dialog} closedby="any">
-  <SvgIcon size={32} alt="Error" fixed>
-    {@html questionMarkSVG}
-  </SvgIcon>
-  This location does not exist or has been deleted.
-  <button onclick={closeDialog}>OK</button>
-</dialog>
 
 <style lang="scss">
   .viewport {
@@ -135,54 +103,5 @@
     display: flex;
     align-items: center;
     justify-content: center;
-  }
-
-  .error-message {
-    all: unset;
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-
-    color: var(--warning);
-    background-color: var(--bg-accent);
-    border-radius: 8px;
-    padding: 16px;
-    box-shadow: var(--box-shadow);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    gap: 16px;
-
-    &::backdrop {
-      background-color: rgba(0, 0, 0, 0.5);
-    }
-
-    @media (prefers-color-scheme: dark) {
-      background-color: var(--bg);
-
-      &::backdrop {
-        background-color: rgba(0, 0, 0, 0.7);
-      }
-    }
-
-    button {
-      color: var(--warning);
-      background-color: var(--warning-bg);
-      border: 1px solid var(--border);
-      border-radius: 4px;
-      padding: 4px 16px;
-      cursor: pointer;
-      font-weight: 500;
-
-      &:hover {
-        background-color: var(--warning-accent-hover);
-      }
-
-      &:active {
-        background-color: var(--warning-accent-active);
-      }
-    }
   }
 </style>
